# merkle: pre‑reserve leaves to prevent reallocs with odd vtx count