### Summary
`ComputeMerkleRoot` duplicates the last hash when the input size is odd:
https://github.com/bitcoin/bitcoin/blob/39b6c139bd6be33699af781f1d71f6fed303d468/src/consensus/merkle.cpp#L54-L56
If the caller gave it a `std::vector` whose capacity equals its size (the common case when the vector was created with `resize()`), that single `push_back` forces a reallocation (doubling its size).
We pay this penalty on every Merkle‑root calculation even though we know the final size in advance.

### Fix
What this PR does:
* call sites: replace `resize(n)` + index assignment with `reserve((n + 1) & ~1ULL)` + `emplace_back(...)`;
* guarantees one extra slot when `vtx.size()` is odd (filled later);
* removes default construction of `uint256` objects we overwrite anyway.

### Validation

The bench was updated to use an even leaf count for simplicity and to focus on hashing speed rather than reallocations.

Asserts (not pushed in this PR): temporary checks confirm that push_back never reallocates and that the coinbase witness hash remains null:

```C++
if (hashes.size() & 1) {
    assert(hashes.size() < hashes.capacity()); // TODO remove
    hashes.push_back(hashes.back());
}
```
and
```C++
leaves.reserve((block.vtx.size() + 1) & ~1ULL); // capacity rounded up to even
leaves.emplace_back();
assert(leaves.back().IsNull()); // TODO remove
```

A full `-reindex-chainstate` up to block **896 408** ran without triggering the asserts.