@yuvicc 2025-06-13
I have thought on changing the benchmark tests as we would want to test the worst case scenario right? or else we could have both case for testing?