@l0rinc 2025-06-15
Update the benchmark to showcase the before/after state better (resembles production code changes), by splitting out the vector copies to the unmetered lambda.
Changed the previous `leaves.reserve((block.vtx.size() + 1) & ~1ULL)` rounding to `leaves.reserve(block.vtx.size() + (block.vtx.size() & 1))`.