@l0rinc 2025-06-25
I've rebased the changed and adjusted the benchmark to be more similar to the other production usages in the first commit, rounded to even in the second (optimization) commit, so that we can realistically measure the speed difference before and after:

% build/bin/bench_bitcoin -filter='MerkleRoot' --min-time=1000

Before 7f620cffebee593e48434cf182cc2fd64a6d76be:

|             ns/leaf |              leaf/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|               45.50 |       21,975,688.66 |    0.0% |      1.10 | `MerkleRoot`
|               45.53 |       21,962,546.91 |    0.1% |      1.10 | `MerkleRoot`
|               45.53 |       21,965,162.52 |    0.0% |      1.10 | `MerkleRoot`

After a73dd45e95f5e909d0950ba7c76225589df74854:
|             ns/leaf |              leaf/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|               45.04 |       22,200,976.95 |    0.1% |      1.10 | `MerkleRoot`
|               44.97 |       22,235,208.25 |    0.1% |      1.10 | `MerkleRoot`
|               45.01 |       22,217,033.81 |    0.1% |      1.10 | `MerkleRoot`