@l0rinc 2025-07-30
Rebased and added a new commit on top for deduplicating the integer rounding. Ready for review again.