@optout21 2025-08-07
LGTM!
Minor comments:
- Maybe add a note to the `ComputeMerkleRoot` declaration, that the `hashes` input is expected to have one extra capacity (in case size is odd). Minor, as this function is not used directly much.
- Even more, it could debug-assert that capacity is even, but this may be a bit of a stretch.